refactor(#334): 툴바 컴포넌트 공통화
